Merge branch 'master' of https://github.com/joshuacroff/spare-change

+++ db/migrate/20150523145830_create_transactions.rb
@@ -1,8 +1,8 @@
 class CreateTransactions < ActiveRecord::Migration
   def change
     create_table :transactions do |t|
-      t.decimal :amount, precision: 8, scale: 2
-      t.decimal :original_amount, precision: 8, scale: 2
+      t.float :amount, scale: 2
+      t.float :original_amount, scale: 2
       t.references :user
     end
   end
